refactor: Empty out infrastructure folder
move router.dart to user/presentation
move network_monitor.dart to core
move deps_resolver.dart to user/presentation
